requirements.txt
streamlit==1.28.1
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
langchain-community==0.0.8
python-dotenv==1.0.0
faiss-cpu==1.8.0
beautifulsoup4==4.12.2
requests==2.31.0
numpy==1.24.3
pandas==2.0.3